Weapon Swaps
Melee Weapons can now swap

+++ Project Y/Assets/Scripts/PlayerScripts/HeavyAttack.cs
@@ -5,7 +5,7 @@ using UnityEngine;
 
 public class HeavyAttackArea : MonoBehaviour
 {
-    private int heavyAttackDamage = 10;
+    private float heavyAttackDamage = 10;
     private void OnTriggerEnter2D(Collider2D collider)
     {
         if (collider.GetComponent<EnemyHealth>() != null)
